+++ CHANGES.md
@@ -3,6 +3,11 @@ The released versions correspond to PyPi releases.
 
 ## Version 4.6.0 (as yet unreleased)
 
+### Fixes
+* fixed handling of alternative path separator in `os.path.split`, 
+  `os.path.splitdrive` and `glob.glob`
+  (see [#632](../../issues/632))
+  
 ## [Version 4.5.1](https://pypi.python.org/pypi/pyfakefs/4.5.1) (2021-08-29)
 This is a bugfix release.